requirements.txt
gradio==4.44.0
huggingface_hub==0.25.1
python-dotenv==1.0.1
